genai-sample-apps/Multimodal Intelligent Response Assistant/requirements.txt
streamlit  
python-dotenv  
Pillow  
streamlit-mic-recorder  
generative-ai-hub-sdk
